requirements.txt
aiofiles==23.2.1
aniso8601==9.0.1
anyio==4.1.0
attrs==23.1.0
azure-core==1.29.5
azure-storage-blob==12.19.0
blinker==1.7.0
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
ci-info==0.3.0
click==8.1.7
configobj==5.0.8
configparser==6.0.0
cryptography==41.0.7
etelemetry==0.3.1
filelock==3.13.1
fitz==0.0.1.dev2
Flask==2.3.3
flask-restx==1.2.0
frontend==0.0.3
future==0.18.3
h11==0.14.0
httplib2==0.22.0
idna==3.6
importlib-resources==6.1.1
isodate==0.6.1
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.17.3
looseversion==1.3.0
lxml==4.9.3
MarkupSafe==2.1.3
networkx==3.2.1
nibabel==5.1.0
nipype==1.8.6
numpy==1.26.2
packaging==23.2
pandas==2.1.3
pathlib==1.0.1
prov==2.0.0
pycparser==2.21
pydot==1.4.2
pyparsing==3.1.1
PyPDF2==3.0.1
pyrsistent==0.20.0
python-dateutil==2.8.2
pytils==0.4.1
pytz==2023.3.post1
pyxnat==1.6
rdflib==7.0.0
requests==2.31.0
scipy==1.11.4
simplejson==3.19.2
six==1.16.0
sniffio==1.3.0
starlette==0.33.0
tools==0.1.9
traits==6.3.2
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.1.0
uvicorn==0.24.0.post1
Werkzeug==2.3.8
